Refactoring Directory
Removed duplicate directory and renamed main directory to dealseeker instead of dealseeker_login as it is not just the login page
