+++ app/src/main/java/com/example/dealseeker/Home.kt
@@ -1,4 +1,4 @@
-package com.example.dealseeker_login
+package com.example.dealseeker
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -13,10 +13,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [PhonesAndGPS.newInstance] factory method to
+ * Use the [Home.newInstance] factory method to
  * create an instance of this fragment.
  */
-class PhonesAndGPS : Fragment(R.layout.fragment_phones_and_g_p_s) {
+class Home : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +34,7 @@ class PhonesAndGPS : Fragment(R.layout.fragment_phones_and_g_p_s) {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_phones_and_g_p_s, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
     }
 
     companion object {
@@ -44,20 +44,16 @@ class PhonesAndGPS : Fragment(R.layout.fragment_phones_and_g_p_s) {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment PhonesAndGPS.
+         * @return A new instance of fragment Home.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            PhonesAndGPS().apply {
+            Home().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
                 }
             }
-
-        fun setOnClickListener(function: () -> Unit) {
-
-        }
     }
 }
