+++ app/src/main/java/com/example/dealseeker/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.dealseeker_login.ui.theme
+package com.example.dealseeker.ui.theme
 
 import androidx.compose.ui.graphics.Color
