+++ app/src/main/java/com/example/dealseeker/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.dealseeker_login.ui.theme
+package com.example.dealseeker.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
