+++ app/src/main/res/layout/activity_main.xml
@@ -30,7 +30,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="com.example.dealseeker_login.PhonesAndGPS"
+        android:name="com.example.dealseeker.PhonesAndGPS"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintBottom_toTopOf="@+id/bottomNavigationView" />
